+++ .env
@@ -4,4 +4,4 @@ POSTGRES_DB=reporting_dashboard_dev
 POSTGRES_USER=reporting_dashboard_role
 POSTGRES_PASSWORD=reportingdashboard123
 DEBUG=join-monster
-ACCESS_TOKEN_SECRET= 4cd7234152590dcfe77e1b6fc52e84f4d30c06fddadd0dd2fb42cbc51fa14b1bb195bbe9d72c9599ba0c6b556f9bd1607a8478be87e5a91b697c74032e0ae7af
+ACCESS_TOKEN_SECRET= "${{ secrets.ACCESS_TOKEN_SECRET }} || 4cd7234152590dcfe77e1b6fc52e84f4d30c06fddadd0dd2fb42cbc51fa14b1bb195bbe9d72c9599ba0c6b556f9bd1607a8478be87e5a91b697c74032e0ae7af"
